separating java-script code on moduls
